requirements.txt
assemblyai
python-dotenv
